Add travis and some renaming
add travis config for CI with docker
change naming for syntax classes
update mount config to include new setting
