+++ .gitignore
@@ -81,3 +81,4 @@ project/plugins/project/
 ###################################################
 ## Project.gitignore
 ###################################################
+.idea
