+++ core/src/main/scala/janstenpickle/scala/syntax/syntax.scala
@@ -8,56 +8,23 @@ import io.circe.syntax._
 import janstenpickle.vault.core.VaultConfig
 import uscala.concurrent.result.AsyncResult
 import uscala.result.Result
-import uscala.result.Result.{Fail, Ok}
 
 import scala.concurrent.{ExecutionContext, Future}
 
-object SyntaxConversion {
+object ConversionSyntax {
   implicit def toResult[A, B](xor: Either[A, B]): Result[A, B] = xor.fold(
     Result.fail, Result.ok
   )
-
-
-  import cats.Applicative
-  import cats.syntax.cartesian._
-
-  type Shit[A] = Result[List[String], A]
-
-  implicit def validatedApplicative: Applicative[Shit] =
-    new Applicative[Shit] {
-      def ap[A, B](f: Shit[A => B])(fa: Shit[A]): Shit[B] =
-        (fa, f) match {
-          case (Ok(a), Ok(fab)) => Ok(fab(a))
-          case (Fail(e1), Fail(e2)) => Fail(e1 ++ e2)
-          case (Ok(_), i@Fail(_)) => i
-          case (i@Fail(_), Ok(_)) => i
-        }
-
-      def pure[A](x: A): Shit[A] = Result.ok(x)
-
-      override def product[A, B](
-        fa: Shit[A],
-        fb: Shit[B]
-      ): Shit[(A, B)] = ap(map(fa)(a => (b: B) => (a, b)))(fb)
-
-      override def map[A, B](fa: Shit[A])(f: A => B): Shit[B] = ap(pure(f))(fa)
-    }
-
-  val shit = (
-    Result.fail[List[String], Int](List("sdfsf")
-  )
-  |@| Result.fail[List[String], Int](List("3423rwfef"))
-  |@| Result.ok[List[String], String]("sdfsf")).map((x, y, z) => (x, y, z))
 }
 
-object SyntaxOption {
+object OptionSyntax {
   implicit class ToTuple[T](opt: Option[T]) {
     def toMap(key: String): Map[String, T] =
       opt.fold[Map[String, T]](Map.empty)(v => Map(key -> v))
   }
 }
 
-object SyntaxAsyncResult {
+object AsyncResultSyntax {
   implicit class FutureToAsyncResult[T](future: Future[T])
   (implicit ec: ExecutionContext) {
     def toAsyncResult: AsyncResult[String, T] = AsyncResult(
@@ -75,7 +42,7 @@ object SyntaxAsyncResult {
     future.toAsyncResult
 }
 
-object SyntaxVaultConfig {
+object VaultConfigSyntax {
 
   final val VaultTokenHeader = "X-Vault-Token"
 
@@ -88,8 +55,8 @@ object SyntaxVaultConfig {
   }
 }
 
-object SyntaxJson {
-  import SyntaxConversion._
+object JsonSyntax {
+  import ConversionSyntax._
 
   implicit class JsonHandler(json: AsyncResult[String, Json]) {
     def extractFromJson[T](jsonPath: HCursor => ACursor = _.downArray)
@@ -103,9 +70,9 @@ object SyntaxJson {
   }
 }
 
-object SyntaxResponse {
-  import SyntaxConversion._
-  import SyntaxJson._
+object ResponseSyntax {
+  import ConversionSyntax._
+  import JsonSyntax._
 
   implicit class ResponseHandler(resp: AsyncResult[String, Response]) {
     def acceptStatusCodes(codes: Int*)
@@ -153,4 +120,3 @@ object SyntaxRequest {
     def delete: Req = req.DELETE
   }
 }
-
