+++ core/src/main/scala/janstenpickle/vault/core/Secrets.scala
@@ -2,8 +2,8 @@ package janstenpickle.vault.core
 
 import com.ning.http.client.Response
 import janstenpickle.scala.syntax.SyntaxRequest._
-import janstenpickle.scala.syntax.SyntaxResponse._
-import janstenpickle.scala.syntax.SyntaxVaultConfig._
+import janstenpickle.scala.syntax.ResponseSyntax._
+import janstenpickle.scala.syntax.VaultConfigSyntax._
 import uscala.concurrent.result.AsyncResult
 import uscala.result.Result
