+++ core/src/main/scala/janstenpickle/vault/core/VaultConfig.scala
@@ -5,9 +5,9 @@ import java.net.URL
 import dispatch.{Req, url}
 import io.circe.generic.auto._
 import io.circe.syntax._
-import janstenpickle.scala.syntax.SyntaxAsyncResult._
+import janstenpickle.scala.syntax.AsyncResultSyntax._
 import janstenpickle.scala.syntax.SyntaxRequest._
-import janstenpickle.scala.syntax.SyntaxResponse._
+import janstenpickle.scala.syntax.ResponseSyntax._
 import uscala.concurrent.result.AsyncResult
 
 import scala.concurrent.ExecutionContext
