+++ manage/src/main/scala/janstenpickle/vault/manage/manage.scala
@@ -3,10 +3,10 @@ package janstenpickle.vault.manage
 import com.ning.http.client.Response
 import io.circe.generic.auto._
 import io.circe.syntax._
-import janstenpickle.scala.syntax.SyntaxOption._
+import janstenpickle.scala.syntax.OptionSyntax._
 import janstenpickle.scala.syntax.SyntaxRequest._
-import janstenpickle.scala.syntax.SyntaxResponse._
-import janstenpickle.scala.syntax.SyntaxVaultConfig._
+import janstenpickle.scala.syntax.ResponseSyntax._
+import janstenpickle.scala.syntax.VaultConfigSyntax._
 import janstenpickle.vault.core.VaultConfig
 import janstenpickle.vault.manage.Model._
 import uscala.concurrent.result.AsyncResult
@@ -99,8 +99,9 @@ object Model {
                    description: Option[String],
                    config: Option[MountConfig])
   case class MountConfig(
-    default_lease_ttl: Int,
-    max_lease_ttl: Int
+    default_lease_ttl: Option[Int],
+    max_lease_ttl: Option[Int],
+    force_no_cache: Option[Boolean]
   )
 
   case class PolicySetting(name: String, rules: Option[String]) {
